refactor: refactor orchestrator to be more consistent with other components
